Move Router to a separate file Rename golem.database.migration.creator to golem.database.migration.create
